Some more legal polish i am not a lawyer i am scared
